Mulsitig plugin customizations, tests adapted

+++ src/conditions/StandardProposalCondition.sol
@@ -12,7 +12,7 @@ import {OptimisticTokenVotingPlugin} from "../OptimisticTokenVotingPlugin.sol";
 /// @notice An abstract contract for non-upgradeable contracts instantiated via the `new` keyword  to inherit from to support customary permissions depending on arbitrary on-chain state.
 contract StandardProposalCondition is ERC165, IPermissionCondition {
     address dao;
-    uint32 minDelay;
+    uint64 minDelay;
 
     error EmptyDao();
     error EmptyDelay();
@@ -22,7 +22,7 @@ contract StandardProposalCondition is ERC165, IPermissionCondition {
      * @param _dao The address of the DAO on which permissions are defined
      * @param _minDelay The minimum amount of seconds to enforce for proposals created
      */
-    constructor(address _dao, uint32 _minDelay) {
+    constructor(address _dao, uint64 _minDelay) {
         if (_dao == address(0)) revert EmptyDao();
         else if (_minDelay == 0) revert EmptyDelay();
